feat: implement services and changes on exception, model and repository

+++ src/main/java/xyz/lorcasdev/model/Item.java
@@ -61,7 +61,7 @@ public class Item {
         return description;
     }
 
-    public double getBaseUnitPrice() {
+    public int getBaseUnitPrice() {
         return baseUnitPrice;
     }
